+++ build.gradle.kts
@@ -1,7 +1,9 @@
 plugins {
+    alias(libs.plugins.kotlin.multiplatform).apply(false)
     alias(libs.plugins.kotlin.android).apply(false)
     alias(libs.plugins.android.lib).apply(false)
     alias(libs.plugins.android.app).apply(false)
+    alias(libs.plugins.compose.multiplatform).apply(false)
     alias(libs.plugins.kotlinx.kover)
     alias(libs.plugins.nexus.publish)
     alias(libs.plugins.dokka)
