+++ gradle.properties
@@ -19,3 +19,7 @@ android.useAndroidX=true
 android.enableJetifier=true
 # Kotlin code style for this project: "official" or "obsolete":
 kotlin.code.style=official
+# KMP
+kotlin.mpp.androidSourceSetLayoutVersion=2
+# Compose Multiplatform
+org.jetbrains.compose.experimental.uikit.enabled=true
