+++ treemap-chart-compose/src/commonMain/kotlin/by/overpass/treemapchart/compose/LocalTreemapChartMeasurer.kt
@@ -1,0 +1,7 @@
+package by.overpass.treemapchart.compose
+
+import androidx.compose.runtime.compositionLocalOf
+import by.overpass.treemapchart.core.measure.TreemapChartMeasurer
+import by.overpass.treemapchart.core.measure.squarified.SquarifiedMeasurer
+
+val LocalTreemapChartMeasurer = compositionLocalOf<TreemapChartMeasurer> { SquarifiedMeasurer() }
