+++ treemap-core/src/commonMain/kotlin/by/overpass/treemapchart/core/tree/Tree.kt
