+++ treemap-example/src/main/java/by/overpass/treemapchart/example/EasyChart.kt
@@ -11,7 +11,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
-import by.overpass.treemapchart.android.TreemapChart
+import by.overpass.treemapchart.compose.TreemapChart
 import by.overpass.treemapchart.core.tree.tree
 
 private val simpleTreeData = tree(10) {
